Move conversion functions to convert.py (#46)
