+++ fluxy/operators/mf.py
@@ -1,7 +1,8 @@
 import numpy as np
 import xarray as xr
 import logging
-from fluxy.operators.select import get_unique_sites, get_site_index, get_variables
+from fluxy.operators.select import get_unique_sites, get_site_index
+from fluxy.operators.convert import get_variables
 
 logger = logging.getLogger(__name__)
